+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to this project will be documented in this file.
 - Message of PopupMonitor
 - Update Advance UI
 - Update delete tip
+- Update toast config
 
 
 ### Fixes